requirements.txt
python-telegram-bot==20.7
openai-whisper==20231117
gfpgan==1.3.8
realesrgan==0.3.0
basicsr==1.4.2
facexlib==0.3.0
ffmpeg-python==0.2.0
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
numpy==1.24.4
